Refactor :Introduce CandidateMapper for clean data transformation
* Improved code readability in the candidate service
* Refactored methods to reduce duplication and improve maintainability

+++ candidate-service/src/main/java/org/dnyanyog/service/CandidateService.java
@@ -11,7 +11,7 @@ public interface CandidateService {
 
  // public CandidateResponse updateCandidate(CandidateRequest request) throws Exception;
 
-  public CandidateResponse deleteByresumeMediaId(long resumeMediaId);
+  public CandidateResponse deleteByresumeMediaId(Integer resumeMediaId);
 
   public List<CandidateResponse> getAllCandidates(CandidateRequest requet);
